Fix another 32-bit system rintf warning

+++ opencog/guile/SchemeSmobAS.cc
@@ -77,7 +77,8 @@ std::string SchemeSmob::as_to_string(const AtomSpace *as)
 #define BUFLEN 120
 	char buff[BUFLEN];
 
-	snprintf(buff, BUFLEN, "#<atomspace %lu>", as->get_uuid());
+	snprintf(buff, BUFLEN, "#<atomspace %lu>",
+		(long unsigned int) as->get_uuid());
 	return buff;
 }
